Retrieve the Naamgeving for a given KvK number

This function retrieves the name information (naamgeving) for a given KvK number (kvkNummer) using the KvK Naamgevingen API.

Usage

kvk_get_naamgeving(kvkNummer, test_environment = FALSE)

Arguments

kvkNummer

A string representing the KvK number for which the name information is requested.

test_environment

A logical value indicating whether to use the test environment (TRUE) or the production environment (FALSE). Defaults to FALSE.

Value

A tibble containing the retrieved name information.

Details

The function also supports the test_environment argument. When set to TRUE, the function will query the KvK test API environment, providing a set of fictitious test data.

The function retrieves data from the KvK Naamgevingen API.

Important: If test_environment = TRUE, no actual API key is required. You will be using test data from the KvK test environment.

Examples

# Examples using the production API (requires API key)
# Retrieve naamgeving for a given KvK number
naamgeving <- kvk_get_naamgeving(kvkNummer = "68750110")
print(naamgeving)
#> # A tibble: 1 × 5
#>   kvkNummer naam     startdatum vestigingen links     
#>   <chr>     <chr>    <chr>      <list>      <list>    
#> 1 68750110  Bikerfix 20170512   <list [1]>  <list [2]>

# Examples using test environment (no API key required)
naamgeving_test <- kvk_get_naamgeving(kvkNummer = "68727720", test_environment = TRUE)
#> ℹ You are using the KvK test environment. Test data will be returned.
print(naamgeving_test)
#> # A tibble: 1 × 7
#>   kvkNummer rsin      statutaireNaam         naam  startdatum vestigingen links 
#>   <chr>     <chr>     <chr>                  <chr> <chr>      <list>      <list>
#> 1 68727720  857567433 Test NV Katrien - Sta… Test… 20160516   <list [1]>  <list>