Retrieve all results from the KvK Search API (up to 1.000 results)

This function automatically paginates through the KvK API to jlajdsflkj l;aj asdlfj collect available results. Due to API limitations, it retrieves a maximum of 1.000 records.

Available arguments are:

kvkNummer: Dutch KVK number. Consists of 8 digits.

rsin: Legal Entities and Partnerships Identification Number.

vestigingsnummer: Branch number. Unique number consisting of 12 digits.

naam: The name under which an establishment or legal entity trades. This can be: active or inactive statutory name, trade name, or name.

straatnaam

huisnummer: Only in combination with postal code.

huisletter: Only in combination with house number. Consists of 1 (alphabetic) character.

postcode: Only in combination with house number.

plaats

postbusnummer: Only allowed in combination with postcode

type: Filter by type: main branch, branch, and/or legal entity. Combine multiple by using ‘&’. E.g. : type=nevenvestiging&type=hoofdvestiging&type=rechtspersoon.

inclusiefInactieveRegistraties: Possible values are "true" or "false". Default value is "false".

Usage

kvk_search(...)

Arguments

...

Named arguments passed to the API query (e.g., naam = "Koudum").

Value

A tibble containing the retrieved results with the following parameters:

kvkNummer: Dutch KVK number. Consists of 8 digits.

rsin: If RSIN is also input.

vestigingsnummer: Branch number. Unique number consisting of 12 digits.

naam: The name under which a branch or legal entity trades. This field contains the (statutory) name or the first trade name. If you searched for a name that is not the statutory or first trade name, we will show the trade name you used for your search query.

adres: List which content indicates whether it's a correspondence or a visiting address, containing the following information:

straatnaam: Street name

huisnummer: If postal code and house number are also input.

huisletter: If postal code and house number are also input.

postcode: If postal code and house number are also input.

plaats: Place

type: Main branch, branch office, and/or legal entity.

actief: Indicates whether a company/establishment/legal entity is registered (Yes) or deregistered (No).

vervallenNaam: Contains the expired trade name or statutory name with which this search result was found.

links: No links will be shown if the deregistration occurred before 22 May 2010 and if "active" is "false". Otherwise a List containing:

Link to the first search result.

Link to Basisprofiel (based on KVK number).

Link to Vestigingsprofiel (based on branch number).

Examples

koudum <- kvk_search(plaats = "Koudum")
#> Error in kvk_search(plaats = "Koudum"): API key is missing. Set it using `Sys.setenv(KVK_API_KEY='your_key')`
koudum
#> Error: object 'koudum' not found