Retrieve all results from the KvK Search API (up to 1.000 results)

This function automatically paginates through the KvK API to collect available results. Due to API limitations, it retrieves a maximum of 1.000 records.

Usage

kvk_search(...)

Arguments

...

Named arguments passed to the API query (e.g., naam = "Koudum"). Available arguments are:

kvkNummer: Dutch KVK number. Consists of 8 digits.

rsin: Legal Entities and Partnerships Identification Number.

vestigingsnummer: Branch number. Unique number consisting of 12 digits.

naam: The name under which an establishment or legal entity trades. This can be: active or inactive statutory name, trade name, or name.

straatnaam

huisnummer: Only in combination with postal code.

huisletter: Only in combination with house number. Consists of 1 (alphabetic) character.

postcode: Only in combination with house number.

plaats

postbusnummer: Only allowed in combination with postcode

type: Filter by type: main branch, branch, and/or legal entity. Combine multiple by using ‘&’. E.g. : type=nevenvestiging&type=hoofdvestiging&type=rechtspersoon.

inclusiefInactieveRegistraties: Possible values are "true" or "false". Default value is "false".

Value

A tibble containing the retrieved results with the following parameters:

kvkNummer: Dutch KVK number. Consists of 8 digits.

rsin: If RSIN is also input.

vestigingsnummer: Branch number. Unique number consisting of 12 digits.

naam: The name under which a branch or legal entity trades. This field contains the (statutory) name or the first trade name. If you searched for a name that is not the statutory or first trade name, we will show the trade name you used for your search query.

adres: List which content indicates whether it's a correspondence or a visiting address, containing the following information:

straatnaam: Street name

huisnummer: If postal code and house number are also input.

huisletter: If postal code and house number are also input.

postcode: If postal code and house number are also input.

plaats: Place

type: Main branch, branch office, and/or legal entity.

actief: Indicates whether a company/establishment/legal entity is registered (Yes) or deregistered (No).

vervallenNaam: Contains the expired trade name or statutory name with which this search result was found.

links: No links will be shown if the deregistration occurred before 22 May 2010 and if "active" is "false". Otherwise a List containing:

Link to the first search result. 2. Link to Basisprofiel (based on KVK number). 3. Link to Vestigingsprofiel (based on branch number).

Examples

koudum <- kvk_search(plaats = "Koudum")
koudum
#> # A tibble: 514 × 6
#>    kvkNummer vestigingsnummer naam                     adres        type  links 
#>    <chr>     <chr>            <chr>                    <list>       <chr> <list>
#>  1 01036576  000007810083     Stichting Gemeenschapsc… <named list> neve… <list>
#>  2 60112891  000045511098     Corriente 't Kofjehúske  <named list> neve… <list>
#>  3 92692966  000032358083     M. de Boer               <named list> neve… <list>
#>  4 40005051  000021251479     Bogerman                 <named list> neve… <list>
#>  5 30168016  000057781184     It Fûgelnêst             <named list> neve… <list>
#>  6 58496505  000027833429     HFK Verhuur              <named list> hoof… <list>
#>  7 62724843  000031708129     De Klink Exploitatie B.… <named list> hoof… <list>
#>  8 01123666  000040288196     Winterberging IJzendoorn <named list> hoof… <list>
#>  9 40001473  000000303062     De Sândobbe              <named list> hoof… <list>
#> 10 01091668  000000678279     Multiservice Beheer en … <named list> hoof… <list>
#> # ℹ 504 more rows

rotterdam <- kvk_search(plaats = "Rotterdam")
#> Warning: API response contains more than 1,000 results. Only the first 1,000 will be retrieved.
rotterdam
#> # A tibble: 1,000 × 6
#>    kvkNummer vestigingsnummer naam                     adres        type  links 
#>    <chr>     <chr>            <chr>                    <list>       <chr> <list>
#>  1 41132991  000023110112     Stichting Albeda         <named list> neve… <list>
#>  2 41132991  000023110597     Stichting Albeda         <named list> neve… <list>
#>  3 41132991  000023110600     Stichting Albeda         <named list> neve… <list>
#>  4 41132991  000023110619     Stichting Albeda         <named list> neve… <list>
#>  5 41132991  000023113456     Stichting Albeda         <named list> neve… <list>
#>  6 24483298  000023336749     Klooster Afrikaander     <named list> neve… <list>
#>  7 24483298  000024590444     Stadsarchief Rotterdam   <named list> neve… <list>
#>  8 24193887  000024452262     Meertens B.V.            <named list> neve… <list>
#>  9 34360247  000022666397     UWV WERKbedrijf Rotterd… <named list> neve… <list>
#> 10 24483298  000022493352     Gemeentewerken Rotterda… <named list> neve… <list>
#> # ℹ 990 more rows